refactor: Complete Phase 1 foundation with TypeScript and modular constants
- Convert core utility files to TypeScript with proper type annotations
- Consolidate duplicate utilities (removed array.js, date.js, price.js, debounce.js)
- Extract constants into organized modular structure:
  - app.ts: Application config, API endpoints, UI settings
  - messages.ts: Error and success message constants
  - validation.ts: Form validation rules and regex patterns
  - stock.ts: Stock status configurations with types
  - admin.ts: Admin interface configuration
  - navigation.ts: Navigation structures with backward compatibility
- Update all import statements to use consolidated utils.ts
- Maintain JSDoc documentation for better IntelliSense
- Preserve backward compatibility through re-exports
- Build verification successful with no breaking changes
